# docs/api-documentation.yml
openapi: 3.0.3
info:
  title: UberFoods API
  version: 1.0.0
  description: Complete API documentation for UberFoods food delivery platform
  contact:
    name: UberFoods API Support
    email: api@uberfoods.com
    url: https://uberfoods.com/api-support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.uberfoods.com/v1
    description: Production server
  - url: https://staging-api.uberfoods.com/v1
    description: Staging server
  - url: http://localhost:3000
    description: Development server
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /restaurants:
    get:
      tags: ["Restaurants"]
      summary: Get restaurants list
      description: Retrieve paginated list of restaurants with filtering options
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
          description: Number of items per page
        - name: cuisine
          in: query
          schema:
            type: string
          description: Filter by cuisine type
        - name: rating
          in: query
          schema:
            type: number
            minimum: 1
            maximum: 5
          description: Minimum rating filter
        - name: deliveryTime
          in: query
          schema:
            type: integer
          description: Maximum delivery time in minutes
        - name: latitude
          in: query
          schema:
            type: number
          description: User latitude for location-based search
        - name: longitude
          in: query
          schema:
            type: number
          description: User longitude for location-based search
      responses:
        200:
          description: Restaurants retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  restaurants:
                    type: array
                    items:
                      $ref: '#/components/schemas/Restaurant'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        400:
          $ref: '#/components/responses/BadRequest'
...
  /orders:
    post:
      tags: ["Orders"]
      summary: Create new order
      description: Place a new food order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - restaurantId
                - items
                - deliveryAddress
              properties:
                restaurantId:
                  type: string
                  example: "rest_123"
                items:
                  type: array
                  items:
                    type: object
                    required:
                      - dishId
                      - quantity
                    properties:
                      dishId:
                        type: string
                        example: "dish_456"
                      quantity:
                        type: integer
                        minimum: 1
                        example: 2
                      specialInstructions:
                        type: string
                        example: "No onions please"
                deliveryAddress:
                  $ref: '#/components/schemas/Address'
                paymentMethod:
                  type: object
                  properties:
                    type:
                      type: string
                      enum: ["card", "paypal", "apple_pay"]
                      example: "card"
                    cardId:
                      type: string
                      example: "card_789"
                tip:
                  type: number
                  minimum: 0
                  example: 5.00
      responses:
        201:
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'

    get:
      tags: ["Orders"]
      summary: Get user orders
      description: Retrieve user's order history
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: status
          in: query
          schema:
            type: string
            enum: ["pending", "confirmed", "preparing", "ready", "picked_up", "delivered", "cancelled"]
      responses:
        200:
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
...
components:
  schemas:
...
    Restaurant:
      type: object
      properties:
        id:
          type: string
          example: "rest_123"
        name:
          type: string
          example: "Mario's Pizza"
        description:
          type: string
          example: "Authentic Italian pizza since 1995"
        cuisine:
          type: string
          example: "Italian"
        rating:
          type: number
          minimum: 0
          maximum: 5
          example: 4.5
        reviewCount:
          type: integer
          example: 1250
        deliveryTime:
          type: integer
          example: 25
        deliveryFee:
          type: number
          example: 2.99
        minimumOrder:
          type: number
          example: 15.00
        isOpen:
          type: boolean
          example: true
        image:
          type: string
          format: uri
          example: "https://cdn.uberfoods.com/restaurants/rest_123.jpg"
        address:
          $ref: '#/components/schemas/Address'
...
    Order:
      type: object
      properties:
        id:
          type: string
          example: "order_123"
        customerId:
          type: string
          example: "user_123"
        restaurantId:
          type: string
          example: "rest_123"
        driverId:
          type: string
          example: "user_456"
        status:
          type: string
          enum: ["pending", "confirmed", "preparing", "ready", "picked_up", "delivered", "cancelled"]
          example: "confirmed"
        items:
          type: array
          items:
            type: object
            properties:
              dishId:
                type: string
                example: "dish_123"
              dishName:
                type: string
                example: "Margherita Pizza"
              quantity:
                type: integer
                example: 2
              price:
                type: number
                example: 12.99
              specialInstructions:
                type: string
                example: "Extra cheese"
        subtotal:
          type: number
          example: 25.98
        deliveryFee:
          type: number
          example: 2.99
        tip:
          type: number
          example: 5.00
        tax:
          type: number
          example: 2.60
        total:
          type: number
          example: 36.57
        deliveryAddress:
          $ref: '#/components/schemas/Address'
        createdAt:
          type: string
          format: date-time
        estimatedDeliveryTime:
          type: string
          format: date-time
...
    Address:
      type: object
      properties:
        street:
          type: string
          example: "123 Main St"
        city:
          type: string
          example: "New York"
        state:
          type: string
          example: "NY"
        zipCode:
          type: string
          example: "10001"
        country:
          type: string
          example: "US"
        latitude:
          type: number
          example: 40.7128
        longitude:
          type: number
          example: -74.0060
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        pages:
          type: integer
          example: 8

  # Error responses
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "VALIDATION_ERROR"
              message:
                type: string
                example: "Invalid input data"
              details:
                type: object
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "UNAUTHORIZED"
              message:
                type: string
                example: "Invalid or missing authentication token"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login

    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for server-to-server communication